requirements.txt
# Rose Medical Pavilion — SEO Agent Dependencies
# pip install -r requirements.txt

anthropic>=0.34.0
pyyaml>=6.0
google-analytics-data>=0.18.0
google-auth>=2.23.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.100.0
requests>=2.31.0
